+++ tracing-futures/src/lib.rs
@@ -194,36 +194,12 @@ mod tests {
 
     #[cfg(feature = "futures-01")]
     use futures::{future, stream, task, Async, Future};
-    #[cfg(feature = "std-future")]
-    use tokio_test::task::MockTask;
     use tracing::{subscriber::with_default, Level};
 
     struct PollN<T, E> {
         and_return: Option<Result<T, E>>,
         finish_at: usize,
         polls: usize,
-    }
-
-    #[cfg(feature = "std-future")]
-    impl<T, E> std::future::Future for PollN<T, E>
-    where
-        T: Unpin,
-        E: Unpin,
-    {
-        type Output = Result<T, E>;
-        fn poll(self: Pin<&mut Self>, lw: &mut Context) -> std::task::Poll<Self::Output> {
-            let this = self.get_mut();
-
-            this.polls += 1;
-            if this.polls == this.finish_at {
-                let value = this.and_return.take().expect("polled after ready");
-
-                std::task::Poll::Ready(value)
-            } else {
-                lw.waker().wake_by_ref();
-                std::task::Poll::Pending
-            }
-        }
     }
 
     #[cfg(feature = "futures-01")]
@@ -258,21 +234,6 @@ mod tests {
                 and_return: Some(Err(())),
                 finish_at,
                 polls: 0,
-            }
-        }
-    }
-
-    #[cfg(feature = "std-future")]
-    fn block_on_future<F>(task: &mut MockTask, future: F) -> F::Output
-    where
-        F: std::future::Future,
-    {
-        let mut future = Box::pin(future);
-
-        loop {
-            match task.poll(&mut future) {
-                std::task::Poll::Ready(v) => break v,
-                _ => {}
             }
         }
     }
@@ -369,44 +330,6 @@ mod tests {
                     });
                 runtime.block_on(Box::new(future)).unwrap();
             })
-        });
-        handle.assert_finished();
-    }
-
-    #[cfg(feature = "std-future")]
-    #[test]
-    fn std_future_enter_exit_is_reasonable() {
-        let (subscriber, handle) = subscriber::mock()
-            .enter(span::mock().named("foo"))
-            .exit(span::mock().named("foo"))
-            .enter(span::mock().named("foo"))
-            .exit(span::mock().named("foo"))
-            .drop_span(span::mock().named("foo"))
-            .done()
-            .run_with_handle();
-        let mut task = MockTask::new();
-        with_default(subscriber, || {
-            let future = PollN::new_ok(2).instrument(span!(Level::TRACE, "foo"));
-            block_on_future(&mut task, future).unwrap();
-        });
-        handle.assert_finished();
-    }
-
-    #[cfg(feature = "std-future")]
-    #[test]
-    fn std_future_error_ends_span() {
-        let (subscriber, handle) = subscriber::mock()
-            .enter(span::mock().named("foo"))
-            .exit(span::mock().named("foo"))
-            .enter(span::mock().named("foo"))
-            .exit(span::mock().named("foo"))
-            .drop_span(span::mock().named("foo"))
-            .done()
-            .run_with_handle();
-        let mut task = MockTask::new();
-        with_default(subscriber, || {
-            let future = PollN::new_err(2).instrument(span!(Level::TRACE, "foo"));
-            block_on_future(&mut task, future).unwrap_err();
         });
         handle.assert_finished();
     }
